requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# frameoverframe/align_image_stack_sequence.py: 17
# frameoverframe/cli/align_image_stack_sequence__main__.py: 17
# frameoverframe/cli/enfuse_batch__main__.py: 21
# frameoverframe/cli/tracer__main__.py: 14
# frameoverframe/enfuse_batch.py: 21
# frameoverframe/img2vid.py: 24
# frameoverframe/tracer.py: 13
Pillow == 7.2.0

# frameoverframe/cli/deflicker__main__.py: 22
docopt == 0.6.2

# frameoverframe/cli/tracer__main__.py: 11,12,13
# frameoverframe/tracer.py: 10,11,12
moviepy == 1.0.3

# frameoverframe/align_image_stack_sequence.py: 19
# frameoverframe/cli/align_image_stack_sequence__main__.py: 19
mulitiprocessingsimple_pkg_uglygus == 0.0.1

# frameoverframe/cli/tracer__main__.py: 10
# frameoverframe/deflicker.py: 4
# frameoverframe/tracer.py: 9
numpy == 1.19.1

# frameoverframe/deflicker.py: 7
progressbar == 2.5

# frameoverframe/align_image_stack_sequence.py: 22
# frameoverframe/autotrace.py: 26
# frameoverframe/cli/align_image_stack_sequence__main__.py: 21
# frameoverframe/img2vid.py: 26
# frameoverframe/utils.py: 11
quotelib == 0.0.1

# frameoverframe/cli/deflicker__main__.py: 23
schema == 0.7.2

# frameoverframe/deflicker.py: 9
scikit_image == 0.17.2

# frameoverframe/deflicker.py: 11
setuptools == 49.2.0
